17: output measured data via websocket

+++ prj_template/include/tpl_net_watchdog.h
@@ -1,6 +1,6 @@
 #ifdef NET_WATCHDOG
 #include <stdint.h>
 
-extern uint8_t tpl_fail;
+extern uint16_t tpl_fail;
 bool tpl_net_watchdog_setup();
 #endif
